# spec/api.yml
openapi: 3.0.3
info:
  title: 診所預約系統 API
  description: |
    診所預約系統的 RESTful API 規格文件。

    每個 API 的 summary 欄位對應 Feature File 中的 step statement。
  version: 1.0.0
  contact:
    name: Clinic Booking System
servers:
  - url: /api
    description: API Base URL
...
paths:
...
  /verification/send:
    post:
      tags:
        - 真人驗證
      summary: 病患請求發送驗證碼
      description: |
        **Feature:** 真人驗證.feature

        **Rules:**
        - 系統必須產生 6 位數驗證碼
        - 驗證碼透過 LINE Messaging API 發送至該用戶
        - 重新發送驗證碼限制為 60 秒內 1 次
      operationId: sendVerificationCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendVerificationRequest'
      responses:
        '200':
          description: 系統產生 6 位數驗證碼，驗證碼透過 LINE Messaging API 發送至該用戶
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendVerificationResponse'
        '429':
          description: 操作失敗 - 距離上次發送未滿 60 秒無法重新發送
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/appointments:
    get:
      tags:
        - 管理員預約
      summary: 取得預約列表
      description: 管理員查看所有預約。
      operationId: getAdminAppointments
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: doctor_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [booked, checked_in, completed, no_show, cancelled]
      responses:
        '200':
          description: 預約列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminAppointmentResponse'

    post:
      tags:
        - 管理員預約
      summary: 管理員新增預約
      description: |
        **Feature:** 管理員手動新增預約.feature

        **Rules:**
        - 時段剩餘分鐘數必須大於等於診療所需分鐘數
        - 病患當日不可有重複預約
        - 預約成功後必須扣除時段分鐘數
        - 預約成功後必須建立預約記錄
        - 若病患有綁定 LINE 則發送通知
      operationId: createAdminAppointment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateAppointmentRequest'
      responses:
        '201':
          description: 預約記錄被建立，時段剩餘分鐘數為 {remainingMinutes}，LINE 訊息發送至 {lineUserId}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAppointmentResponse'
        '400':
          description: 操作失敗 - 時段剩餘分鐘數不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 操作失敗 - 病患當日已有預約
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================
  # 管理員編輯預約.feature
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
...
    SendVerificationRequest:
      type: object
      required:
        - line_user_id
      properties:
        line_user_id:
          type: string
    SendVerificationResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
...
    AppointmentStatus:
      type: string
      enum: [booked, checked_in, completed, no_show, cancelled]
      description: |
        預約狀態：
        - booked: 已預約
        - checked_in: 已報到
        - completed: 已完成
        - no_show: 未報到
        - cancelled: 已取消
...
    AdminCreateAppointmentRequest:
      type: object
      required:
        - patient_id
        - slot_id
        - treatment_id
      properties:
        patient_id:
          type: string
        slot_id:
          type: string
        treatment_id:
          type: string
...
    AdminAppointmentResponse:
      type: object
      properties:
        id:
          type: string
        patient_id:
          type: string
        patient_name:
          type: string
        patient_phone:
          type: string
        patient_line_user_id:
          type: string
        slot_id:
          type: string
        treatment_id:
          type: string
        treatment_name:
          type: string
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        status_label:
          type: string
        appointment_date:
          type: string
          format: date
        slot_start_time:
          type: string
        slot_end_time:
          type: string
        doctor_id:
          type: string
        doctor_name:
          type: string
        cancel_reason:
          type: string
        updated_by:
          type: string
        updated_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time

    # ============================================
    # 管理員帳號
    # ============================================
...